pyproject.toml
[project]
name = "expense-bot"
version = "0.1.0"
description = "Expense tracking telegram bot"
authors = [
    {name = "nicktanqy",email = "nicktanqy@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot (>=22.0,<23.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
